chat template: the conversation template follows the loaded model, not the branch
A 0731 artifact primed with V4.1's template answered a tool request with a DSML
'invoke' and NO block opener, then drifted, so tool-call-quality (a gating unit
test) failed on the merged tree while dev's binary passed it on the same
artifact.  The two families' templates differ in more than the tag spelling --
where the tool schemas sit, the tools-prompt text, how consecutive user-side
messages merge, whether a mid-conversation system message is an in-place System
token, which assistant turns replay reasoning, and how non-string argument
values are encoded -- so the template is a per-profile FACT, like
pulsar_shape::chat_system_marker on the engine's side.

  * pulsar_dsml.h/.cpp: V4's tag spelling is back as rows 2/3 of the table (the
    parsers loop over all rows, so both families' output parses either way) and
    pulsar_dsml_canonical(v41) names the row a renderer WRITES.
  * pulsar_engine_chat_v41() publishes the one fact the server needs.
  * request::chat_v41 carries it from the parser to every writer -- the renderer,
    the forced-tool prefill, the DSML writer and the KV-key suffix builders --
    so the live KV and the replay cannot disagree about which template produced
    the bytes.
  * prompt_render.cpp forks the eleven differing rules; every new parameter
    defaults to V4.1, so the renderer gate (which grades the default profile
    against the V4.1 reference) renders byte-identically.
  * kv_cache's tool-block scanner accepts every spelling.

+++ tests/pulsar_test.cpp
@@ -1108,7 +1108,7 @@ static void test_tool_call_quality_one(void) {
         /* Print WHAT the model produced: "no tool call" alone cannot tell a
          * rendering difference from a decode one. */
         fprintf(stderr, "pulsar-test: tool-call-quality generated %d BYTES:\n%.*s\n",
-                text.len, text.len, text.ptr ? text.ptr : "(empty)");
+                (int)text.len, (int)text.len, text.ptr ? text.ptr : "(empty)");
     }
     TEST_ASSERT(calls.len > 0);
     TEST_ASSERT(calls.len > 0 && !strcmp(calls.v[0].name, "list_files"));
